Refactor/trainer split (#930)
* refactor(training): extract shared helpers from hv_train_network into training package

Move free functions out of hv_train_network.py into a new musubi_tuner.training
subpackage so that each file matches its contents by name:

- training/accelerator_setup.py: clean_memory_on_device, collator_class, prepare_accelerator
- training/sampling_prompts.py:  line_to_prompt_dict, load_prompts, should_sample_images
- training/timesteps.py:         compute_density_for_timestep_sampling, get_sigmas, compute_loss_weighting_for_sd3
- training/parser_common.py:     setup_parser_common, read_config_from_file

hv_train_network.py re-exports these symbols so that existing imports from
wan_train_network.py and other architecture-specific scripts keep working
unchanged. NetworkTrainer itself is left in place for a follow-up commit.

hv_train_network.py: 3096 -> 2094 lines.

* refactor(training): move NetworkTrainer to training.trainer_base

Split the NetworkTrainer class out of hv_train_network.py into its own module
so the file name matches its contents. hv_train_network.py is now just the
HunyuanVideo-specific entry point (hv_setup_parser + main) plus re-exports
for backward compatibility with architecture-specific training scripts that
import NetworkTrainer / helpers from musubi_tuner.hv_train_network.

SS_METADATA_* constants move with NetworkTrainer (it uses them) and are
re-exported from hv_train_network.py.

Fix sys.path fallback for --network_module short names: __file__ now lives
under musubi_tuner/training/, so step one extra level up to keep
`--network_module networks.lora` resolving to musubi_tuner/networks/lora.py.

hv_train_network.py: 2094 -> 87 lines.
training/trainer_base.py: new, 2061 lines.

* refactor(training): split NetworkTrainer.train() into phase methods

Decompose the 693-line train() into private per-phase methods so the top
of the method reads as a table of contents:

- _validate_args_and_init
- _init_session
- _build_dataset
- _prepare_accelerator_and_dtypes
- _prepare_sampling
- _load_dit_and_swap
- _build_network
- _build_optimizer_and_dataloader
- _prepare_with_accelerator
- _register_hooks_and_resume
- _run_training_loop

No behavior change: state flows through return tuples and arguments;
existing self attributes (blocks_to_swap, num_timestep_buckets) are kept.
Verified with --help and by importing all *_train_network.py subclasses;
loss curves over 20 steps on Z-Image match main/Step 1 within numerical
error.

* refactor(training): split setup_parser_common into category helpers

Decompose the 690-line setup_parser_common into _add_*_args helpers
grouped by CLI category (general, attention, compile/dynamo, training,
logging, ddp, sampling, optimizer, lr_scheduler, memory, timestep,
network, save_load, metadata, huggingface, model). setup_parser_common
now reads as a table of contents.

No behavioral change: the resulting parser's argument set is identical
(verified by sorted argument diff). Only --help display order differs —
dynamo args now sit next to compile args instead of after --fp8_base.
The nested int_or_float is lifted to module scope as _int_or_float.

* fix(training): add __all__ to hv_train_network and remove unused imports from trainer_base

hv_train_network re-exports symbols for backward compatibility with
architecture-specific scripts. Adding __all__ suppresses ruff F401
warnings while documenting the intentional re-exports. Also remove
11 unused imports from trainer_base left over from the refactoring.


---------
